Moved enableToolbarKeyboardFocus ToolbarView utils to a separate module.
